Refactor some navigation to improve reusability and accessibility (#541)
* Refactor some navigation to improve reusability and accessibility

Why: we were using suboptimal patterns to create some header navigation
components. We would like to use better patterns to improve
accessibility, UX, and reusability in the codebase.

This commit:
- Updates some navigational elements to use React Navigation
to render them instead of rendering them within the component itself
- Removes the ModalStack and moves all the screens to the MainNavigator

* WIP

* Add back modalstack

* Delete ModalStack

* Navbar Updates (#548)

#### Why:
Several screens which are "pushed" to the nav stack rather than "presented" modally needed to be updated to display the navbar with a `back` button rather than hide the navbar and display an `X` icon

#### This commit:
This commit updates the relevant screens, and also displays the default `Back` label on iOS (to increase the tappable area of this button, which can be difficult to reach on larger iOS devices)

![Oct-11-2020 11-29-48](https://user-images.githubusercontent.com/2637355/95682769-243e9780-0bb5-11eb-9b5f-e25f44415cc2.gif)
