+++ src/AffectedUserFlow/CodeInput/EnableExposureNotifications.spec.tsx
@@ -1,7 +1,6 @@
 import React from "react"
 import { openAppSettings } from "../../gaen/nativeModule"
 import { render, fireEvent } from "@testing-library/react-native"
-import { useNavigation } from "@react-navigation/native"
 
 import EnableExposureNotifications from "./EnableExposureNotifications"
 
@@ -17,16 +16,6 @@ describe("EnableExposureNotifications", () => {
     ).toBeDefined()
     expect(getByText("Enable exposure notifications to continue")).toBeDefined()
   })
-
-  it("navigates to the home screen if users cancels", () => {
-    const navigateSpy = jest.fn()
-    ;(useNavigation as jest.Mock).mockReturnValue({ navigate: navigateSpy })
-    const { getByLabelText } = render(<EnableExposureNotifications />)
-
-    fireEvent.press(getByLabelText("Cancel"))
-    expect(navigateSpy).toHaveBeenCalledWith("Home")
-  })
-
   it("takes user to system settings if decide to activate notifications", () => {
     const { getByLabelText } = render(<EnableExposureNotifications />)
