+++ src/navigation/MainNavigator.tsx
@@ -2,6 +2,7 @@ import React, { FunctionComponent, useRef } from "react"
 import {
   createStackNavigator,
   TransitionPresets,
+  HeaderStyleInterpolators,
 } from "@react-navigation/stack"
 import { Platform } from "react-native"
 import {
@@ -9,21 +10,38 @@ import {
   NavigationContainer,
   NavigationContainerRef,
 } from "@react-navigation/native"
+import { useTranslation } from "react-i18next"
 
+import { ModalStackScreens, HomeStackScreens } from "./index"
 import { useOnboardingContext } from "../OnboardingContext"
+import { useAnalyticsContext } from "../AnalyticsContext"
 import { WelcomeStackScreens, Stacks } from "./index"
 import MainTabNavigator from "./MainTabNavigator"
 import HowItWorksStack from "./HowItWorksStack"
 import ActivationStack from "./ActivationStack"
 import SettingsStack from "./SettingsStack"
-import ModalStack from "./ModalStack"
 import Welcome from "../Welcome"
-import { useAnalyticsContext } from "../AnalyticsContext"
+import LanguageSelection from "../modals/LanguageSelection"
+import ProtectPrivacy from "../modals/ProtectPrivacy"
+import AffectedUserStack from "../AffectedUserFlow/"
+import AnonymizedDataConsentScreen from "../modals/AnonymizedDataConsentScreen"
+import SelfAssessmentStack from "./SelfAssessmentStack"
+import ExposureDetectionStatus from "../Home/ExposureDetectionStatus"
+import BluetoothInfo from "../Home/BluetoothInfo"
+import ExposureNotificationsInfo from "../Home/ExposureNotificationsInfo"
+import LocationInfo from "../Home/LocationInfo"
+import CallbackStack from "./CallbackStack"
+import CovidDataDashboard from "../CovidDataDashboard/CovidDataDashboard"
+import { applyModalHeader } from "./ModalHeader"
+import { applyHeaderLeftBackButton } from "../navigation/HeaderLeftBackButton"
+
+import { Headers } from "../styles"
 
 const Stack = createStackNavigator()
 
 const defaultScreenOptions = {
   headerShown: false,
+  headerStyleInterpolator: HeaderStyleInterpolators.forNoAnimation,
 }
 
 const settingsStackTransitionPreset = Platform.select({
@@ -41,6 +59,7 @@ const linking: LinkingOptions = {
 }
 
 const MainNavigator: FunctionComponent = () => {
+  const { t } = useTranslation()
   const { isOnboardingComplete } = useOnboardingContext()
   const { trackScreenView } = useAnalyticsContext()
   const navigationRef = useRef<NavigationContainerRef>(null)
@@ -68,21 +87,14 @@ const MainNavigator: FunctionComponent = () => {
       ref={navigationRef}
       onStateChange={trackPageView}
     >
-      <Stack.Navigator>
+      <Stack.Navigator headerMode="screen" screenOptions={defaultScreenOptions}>
         {isOnboardingComplete ? (
           <>
-            <Stack.Screen
-              name={"App"}
-              component={MainTabNavigator}
-              options={defaultScreenOptions}
-            />
+            <Stack.Screen name={"App"} component={MainTabNavigator} />
             <Stack.Screen
               name={Stacks.Settings}
               component={SettingsStack}
-              options={{
-                ...settingsStackTransitionPreset,
-                headerShown: false,
-              }}
+              options={settingsStackTransitionPreset}
             />
           </>
         ) : (
@@ -90,12 +102,8 @@ const MainNavigator: FunctionComponent = () => {
             <Stack.Screen
               name={WelcomeStackScreens.Welcome}
               component={Welcome}
-              options={defaultScreenOptions}
             />
-            <Stack.Screen
-              name={Stacks.HowItWorks}
-              options={defaultScreenOptions}
-            >
+            <Stack.Screen name={Stacks.HowItWorks}>
               {(props) => (
                 <HowItWorksStack
                   {...props}
@@ -114,10 +122,108 @@ const MainNavigator: FunctionComponent = () => {
           </>
         )}
         <Stack.Screen
-          name={Stacks.Modal}
-          component={ModalStack}
+          name={ModalStackScreens.LanguageSelection}
+          component={LanguageSelection}
           options={{
-            headerShown: false,
+            ...TransitionPresets.ModalTransition,
+            headerShown: true,
+            header: applyModalHeader(t("screen_titles.select_language")),
+          }}
+        />
+        <Stack.Screen
+          name={ModalStackScreens.ProtectPrivacy}
+          component={ProtectPrivacy}
+          options={{
+            ...TransitionPresets.ModalTransition,
+            headerShown: true,
+            header: applyModalHeader(t("screen_titles.protect_privacy")),
+          }}
+        />
+        <Stack.Screen
+          name={Stacks.AffectedUserStack}
+          component={AffectedUserStack}
+        />
+        <Stack.Screen name={ModalStackScreens.HowItWorksReviewFromSettings}>
+          {(props) => (
+            <HowItWorksStack {...props} destinationOnSkip={Stacks.Settings} />
+          )}
+        </Stack.Screen>
+        <Stack.Screen name={ModalStackScreens.HowItWorksReviewFromConnect}>
+          {(props) => (
+            <HowItWorksStack {...props} destinationOnSkip={Stacks.Connect} />
+          )}
+        </Stack.Screen>
+        <Stack.Screen
+          name={ModalStackScreens.AnonymizedDataConsent}
+          component={AnonymizedDataConsentScreen}
+        />
+        <Stack.Screen
+          name={ModalStackScreens.SelfAssessmentFromExposureDetails}
+        >
+          {(props) => {
+            return (
+              <SelfAssessmentStack
+                {...props}
+                destinationOnCancel={Stacks.ExposureHistoryFlow}
+              />
+            )
+          }}
+        </Stack.Screen>
+        <Stack.Screen name={ModalStackScreens.SelfAssessmentFromHome}>
+          {(props) => {
+            return (
+              <SelfAssessmentStack
+                {...props}
+                destinationOnCancel={Stacks.Home}
+              />
+            )
+          }}
+        </Stack.Screen>
+        <Stack.Screen
+          name={HomeStackScreens.ExposureDetectionStatus}
+          component={ExposureDetectionStatus}
+          options={{
+            ...Headers.headerMinimalOptions,
+            headerLeft: applyHeaderLeftBackButton(),
+          }}
+        />
+        <Stack.Screen
+          name={HomeStackScreens.BluetoothInfo}
+          component={BluetoothInfo}
+          options={{
+            ...TransitionPresets.ModalTransition,
+            headerShown: true,
+            header: applyModalHeader(t("screen_titles.bluetooth")),
+          }}
+        />
+        <Stack.Screen
+          name={HomeStackScreens.ExposureNotificationsInfo}
+          component={ExposureNotificationsInfo}
+          options={{
+            ...TransitionPresets.ModalTransition,
+            headerShown: true,
+            header: applyModalHeader(t("screen_titles.exposure_notifications")),
+          }}
+        />
+        <Stack.Screen
+          name={HomeStackScreens.LocationInfo}
+          component={LocationInfo}
+          options={{
+            ...TransitionPresets.ModalTransition,
+            headerShown: true,
+            header: applyModalHeader(t("screen_titles.location")),
+          }}
+        />
+        <Stack.Screen
+          name={ModalStackScreens.CallbackStack}
+          component={CallbackStack}
+        />
+        <Stack.Screen
+          name={HomeStackScreens.CovidDataDashboard}
+          component={CovidDataDashboard}
+          options={{
+            ...Headers.headerMinimalOptions,
+            headerLeft: applyHeaderLeftBackButton(),
           }}
         />
       </Stack.Navigator>
