+++ src/navigation/index.ts
@@ -213,7 +213,6 @@ export type Stack =
   | "AffectedUserStack"
   | "Connect"
   | "ExposureHistoryFlow"
-  | "Modal"
   | "HowItWorks"
   | "Settings"
   | "Home"
@@ -224,7 +223,6 @@ export const Stacks: { [key in Stack]: Stack } = {
   AffectedUserStack: "AffectedUserStack",
   Connect: "Connect",
   ExposureHistoryFlow: "ExposureHistoryFlow",
-  Modal: "Modal",
   HowItWorks: "HowItWorks",
   Settings: "Settings",
   Home: "Home",
